requirements.txt
astroid==2.12.12
autopep8==1.7.0
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
dill==0.3.5.1
distlib==0.3.6
filelock==3.8.0
identify==2.5.6
idna==3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
nodeenv==1.7.0
platformdirs==2.5.2
pre-commit==2.20.0
pycodestyle==2.9.1
pylint==2.15.4
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.5
wrapt==1.14.1
